chore: move chart legend component to common charts
