+++ src/components/pages/state/race-ethnicity/charts-section.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styles from './charts.module.scss'
-import ChartLegend from './chart-legend'
+import ChartLegend from '~components/charts/chart-legend'
 
 const ChartsSection = ({
   title,
@@ -10,19 +10,32 @@ const ChartsSection = ({
   setSelectedItem,
   legendColors,
   legendRef,
-}) => (
-  <div className={styles.chartSection}>
-    <h3 className={styles.chartSectionTitle}>{title}</h3>
-    {children}
-    <div ref={legendRef}>
-      <ChartLegend
-        legendColors={legendColors}
-        categories={legendCategories}
-        selectedItem={selectedItem}
-        setSelectedItem={setSelectedItem}
-      />
+}) => {
+  const categoryNames = {
+    White: 'White',
+    Black: 'Black or African American',
+    LatinX: 'Hispanic or Latino',
+    Asian: 'Asian',
+    AIAN: 'American Indian and Alaskan Native',
+    NHPI: 'Native Hawaiian and Other Pacific Islander',
+    Ethnicity_Hispanic: 'Hispanic or Latino',
+    Ethnicity_NonHispanic: 'Not Hispanic or Latino',
+  }
+  return (
+    <div className={styles.chartSection}>
+      <h3 className={styles.chartSectionTitle}>{title}</h3>
+      {children}
+      <div ref={legendRef}>
+        <ChartLegend
+          legendColors={legendColors}
+          categories={legendCategories}
+          selectedItem={selectedItem}
+          setSelectedItem={setSelectedItem}
+          legendNames={categoryNames}
+        />
+      </div>
     </div>
-  </div>
-)
+  )
+}
 
 export default ChartsSection
